--- middleware.ts
@@ -1,40 +1,0 @@
-import { createServerClient } from '@supabase/ssr'
-import { NextResponse, type NextRequest } from 'next/server'
-import { supabaseConfig } from '@/utils/supabase/config'
-
-export async function middleware(request: NextRequest) {
-  let supabaseResponse = NextResponse.next({
-    request,
-  })
-
-  const supabase = createServerClient(supabaseConfig.url, supabaseConfig.key, {
-    cookies: {
-      getAll() {
-        return request.cookies.getAll()
-      },
-      setAll(cookiesToSet) {
-        cookiesToSet.forEach(({ name, value, options }) =>
-          request.cookies.set(name, value),
-        )
-        supabaseResponse = NextResponse.next({
-          request,
-        })
-        cookiesToSet.forEach(({ name, value, options }) =>
-          supabaseResponse.cookies.set(name, value, options),
-        )
-      },
-    },
-  })
-
-  // 중요: 이 함수를 호출해야 세션이 만료되었을 때 자동으로 갱신됩니다.
-  await supabase.auth.getUser()
-
-  return supabaseResponse
-}
-
-export const config = {
-  matcher: [
-    // 아래 경로를 제외한 모든 경로에서 미들웨어 실행
-    '/((?!_next/static|_next/image|favicon.ico|.*\\.(?:svg|png|jpg|jpeg|gif|webp)$).*)',
-  ],
-}
